move response to azurecore
