+++ AzureData/Source/AzureData.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import AzureCore
 
 public class AzureData {
