+++ AzureData/Source/AzureDataExtensions.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import AzureCore
 
 // MARK: -
