+++ AzureData/Source/ResponseExtensions.swift
@@ -1,5 +1,5 @@
 //
-//  Response.swift
+//  ResponseExtensions.swift
 //  AzureData
 //
 //  Copyright (c) Microsoft Corporation. All rights reserved.
@@ -9,124 +9,22 @@
 import Foundation
 import AzureCore
 
-public struct Response<T> {
-    
-    public let request: URLRequest?
-    
-    public let response: HTTPURLResponse?
-    
-    public let data: Data?
-    
-    public let result: Result<T>
-    
-    public var error: Error? { return result.error }
-    
-    public var resource: T?  { return result.resource }
-    
-    public var fromCache: Bool = false
-
-    public lazy var metadata: ResponseMetadata? = response.flatMap { ResponseMetadata(for: $0) }
-
-    public init(request: URLRequest?, data: Data?, response: HTTPURLResponse?, result: Result<T>, fromCache: Bool = false) {
-        self.request = request
-        self.data = data
-        self.response = response
-        self.result = result
-        self.fromCache = fromCache
-    }
-
-    public init (_ resource: T, fromCache: Bool = false) {
-        self.init(request: nil, data: nil, response: nil, result: .success(resource), fromCache: fromCache)
-    }
-    
-    public init (_ error: Error, fromCache: Bool = false) {
-        self.init(request: nil, data: nil, response: nil, result: .failure(error), fromCache: fromCache)
-    }
-}
-
-
-public enum Result<T> {
-    case success(T)
-    case failure(Error)
-    
-    public var isSuccess: Bool {
-        switch self {
-        case .success: return true
-        case .failure: return false
-        }
-    }
-    
-    public var isFailure: Bool { return !isSuccess }
-    
-    public var resource: T? {
-        switch self {
-        case .success(let resource): return resource
-        case .failure: return nil
-        }
-    }
-    
-    public var error: Error? {
-        switch self {
-        case .success: return nil
-        case .failure(let error): return error
-        }
-    }
-}
-
-
-extension Response where T:CodableResource {
-
-    public init(request: URLRequest?, data: Data?, response: HTTPURLResponse?, result: Result<T>) {
-        self.request = request
-        self.data = data
-        self.response = response
-        self.result = result
-    }
-}
-
 extension Response {
-    public func map<U>(_ transform: (T) throws -> U) -> Response<U> {
-        return Response<U>(
-            request: request,
-            data: data,
-            response: response,
-            result: result.map(transform),
-            fromCache: fromCache
-        )
-    }
-}
-
-extension Result {
-    public func map<U>(_ transform: (T) throws -> U) -> Result<U> {
-        do {
-            return try isSuccess ? Result<U>.success(transform(resource!)) : Result<U>.failure(error!)
-        } catch {
-            return Result<U>.failure(error)
+    
+    public mutating func responseMetadata() -> ResponseMetadata? {
+        
+        if let data = metadata as? ResponseMetadata {
+            return data
         }
+        
+        let newMetadata = response.flatMap { ResponseMetadata(for: $0) }
+        
+        metadata = newMetadata
+        
+        return newMetadata
     }
 }
 
-// MARK: - CustomStringConvertible
-
-extension Result: CustomStringConvertible {
-    public var description: String {
-        switch self {
-        case .success: return "✅ SUCCESS"
-        case .failure: return "❌ FAILURE"
-        }
-    }
-}
-
-// MARK: - CustomDebugStringConvertible
-
-extension Result: CustomDebugStringConvertible {
-    public var debugDescription: String {
-        switch self {
-        case .success(let value): return "✅ SUCCESS: \(value)"
-        case .failure(let error): return "❌ FAILURE: \(error)"
-        }
-    }
-}
 
 // MARK: - Pagination support
